general: Use colors for `AC_MSG_RESULT`
Color mapping (first word):

green: 'yes', 'ok'
red:   'no', 'failed', 'unsupported'
blue:  *

If the argument is a constant string (i.e., a shell literal) mapping will
be applied at autoconf time, else a AS_CASE is used. In the latter, we
match only 'yes[, ]*' and 'no[, ]*' if it is not a single word.

`AC_MSG_NOTICE` is now mapped to `AS_INFO`, which colors blue.